+++ src/components/Sidebar/ThemeToggler.tsx
@@ -2,7 +2,7 @@ import { ToggleButton, ToggleButtonGroup } from '@mui/material';
 import React from 'react';
 import DarkModeIcon from '@mui/icons-material/DarkMode';
 import LightModeIcon from '@mui/icons-material/LightMode';
-import { THEMES } from './InfoSection';
+import { THEMES } from 'types/theme';
 interface Iprops {
     theme: THEMES;
     setTheme: (theme: THEMES) => void;
