+++ src/types/theme/index.ts
@@ -1,0 +1,4 @@
+export enum THEMES {
+    LIGHT,
+    DARK,
+}
